Added some base assets and code
added standard prefabs with corresponding behavior scripts. Made a Scene
to work with in unity. coded the TankBehavior.TakeDamage() method using
LoL attributes.

- used PolygonCollider2D as the colliders for tanks and terrains. We
need this for terrain, but maybe not for tanks. It might be better to
use a box collider for tanks since we only really need the bottom of the
tank to collide.

-Lucas